Changes to saving/removing models

+++ utils/train_centroid_regressor.py
@@ -143,7 +143,7 @@ for epoch in range(start_epoch, opt.nepoch):
         torch.save(regressor.state_dict(), '%s/cls_model_%d.pth' % (opt.outf, epoch))
 
     # Only keep every 10th
-    if (epoch - 1) % 10 != 0:
+    if epoch > 0 and (epoch - 1) % 10 != 0:
         os.remove('%s/cls_model_%d.pth' % (opt.outf, epoch - 1))
 
 total_correct = 0
